fix benchmarks when run together
